made changes to CategoriesController, TagsController, created migration for users_table, created UsersTableSeeder for database seeding

+++ app/Http/Controllers/TagsController.php
@@ -94,6 +94,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts()->count() > 0)
+        {
+            session()->flash("error", "Tag cannot be deleted -- it has posts.");
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash("success", "Tag deleted successfully..");
